Parallelize sampling (v7 API exclusively)
